coredumpctl: move print_info() and print_entry() to coredumpctl-info.c
